refactor(backend): move files to more suitable packages
